// apps/web-app/src/pages/SuperAdmin/Shared.d.ts
import React from 'react';
export declare const CardStat: React.FC<{
    label: string;
    value: string;
    valueClass?: string;
}>;
export declare const DetailGrid: React.FC<{
    title: string;
    rows: Array<{
        label: string;
        value: string;
    }>;
}>;
//# sourceMappingURL=Shared.d.ts.map